Clean up app delegate
